backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.3
pydantic-settings>=2.2.1
aiohttp>=3.9.3

# Data Processing
pandas>=2.2.3
numpy>=1.26.4
pyarrow>=15.0.0
json5>=0.9.14

# Database and Templates
sqlalchemy>=2.0.26
jinja2>=3.1.3
motor>=3.3.2  # MongoDB async driver
pymongo>=4.6.1

# API and Environment
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.26.0  # Moved from Development as it's needed for async HTTP calls
google-generativeai>=0.3.1  # Google Generative AI (Gemini) SDK

# Development
black>=24.1.1
pytest>=7.4.4
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Error Tracking
sentry-sdk[fastapi]>=1.39.1